requirements.txt
  
twilio
sendgrid
python-dotenv
python-telegram-bot


pydantic
bson
pymongo

transformers 
torch